#11 Convert classmethods to functions (from_*)
